requirements.txt
certifi==2020.12.5
chardet==4.0.0
idna==2.10
MLB-StatsAPI==1.0.1
PySide6==6.0.1
requests==2.25.1
shiboken6==6.0.1
urllib3==1.26.3
